requirements.txt
redis
torch
torchvision
python-dotenv
sentence_transformers
annoy
scipy
numpy
pandas
